Unexpected swact after rebooting standby controller
In AIO-DX, 2 controllers sometimes require longer than expected
(2500ms) to sync the status of the service groups.
The fix here is to extend the wait interval to 5000ms in order to give
enough time for the controllers to finish serivce groups status
exchange.
The extra of wait time impacts only the 1st controller reboot.
In a normal scenario, the 2nd controller boots up the service
scheduling will start immediatelly after the service group status
exchange completed.

Closes-Bug: #24230

Change-Id: Ibc49df8278ec0a0b319691ae9e82e9318a35a2f6
Signed-off-by: Bin Qian <bin.qian@windriver.com>