stx-ha: API ref doc content added.
I added the api-ref-smapi-v1-ha.rst file, which is content for
the API reference manual to the api-ref/source dir. This
represents the converted old-style files to the newer
OpenStack supported RST files.

I updated the index.rst file to include the new .rst file so
that the api-ref document can build.

Provided info for API Details response.

Change-Id: Iae5b7297d27621b7862830e402109ecfa7faba09
Signed-off-by: Scott Rifenbark <srifenbark@gmail.com>