Add DevStack support
This builds, installs and cleans some of the SM components, plus adds
a DevStack job and a build.sh script to build components outside DevStack.

Also some interesting work-arounds were required to build sm-common and sm-db,
need to sort out the differences between CentOS 7.5 and Ubuntu Xenial.

Fixed a couple of occurrances of this error:
  sm_db_service_domain_neighbors.c:311:25: error: ISO C++ forbids comparison between
  pointer and integer [-Werror=permissive]
     if( '\0' != record->orchestration )
                        ^~~~~~~~~~~~~

Change-Id: I33c44a764295e8e0a7ae4771ea02e950f06bf2a2
Signed-off-by: Dean Troyer <dtroyer@gmail.com>