Remove ceph-mon pid file from SM monitoring
This commit fixes a racing issue between SM's PID monitoring and 'worker'
puppet manifests on AIO-DX triggering a restart of ceph-mon.

After a reboot on AIO DX, SM takes control of ceph-mon monitoring after
'controller' puppet manifests finish applying. As part of this, SM monitors
ceph-mon pid files for changes (through inotify) and also periodically
executes '/etc/init.d/ceph status mon.controller' for a more advanced
monitoring.

After the 'controller' puppet manifest finish applying a 2nd set is executed:
the 'worker' puppet manifests. This set triggers a restart of ceph-mon
through /etc/init.d/ceph restartthat has two steps: 'stop' in which ceph-mon
is stopped, and 'start' in which it is restarted.

In the first step, stopping ceph-mon leads to the removal of the PID file.
This is promptly detected by SM which immediately triggers a start of ceph-mon
taht creates a new pid file. Problem is that ceph-mon was already in a restart,
and at the end of the 'stop' step the init script cleans up the new pid file
instead of the old.

This leads to controllers swacting a couple of times before the system
gets rid of the rogue process.

This commit removes SM's monitoring of the PID file as relying on the more
advanced monitoring implemented in '/etc/init.d/ceph status mon.controller'
is enough.

Change-Id: I998bc8083720425bdca10c6876e1c2ed99235bfa
Story: 2002844
Task: 29214
Signed-off-by: Ovidiu Poncea <Ovidiu.Poncea@windriver.com>