Add SM DB entries for monitor for containerized mariadb
For the AIO-DX configuration we have a problem where we can't run with
a single mariadb replica and we can't run with two mariadb replicas and
a garbd daemon like we do on the full configuration.  (See the commit
message for the Depends-On commit for more info.)

Accordingly, add entries for the database monitor into the SM database.

Change-Id: Ibd28fb99eebb8b72a205d2626e49c4e2db4f01dd
Story: 2004712
Task: 30410
Signed-off-by: Chris Friesen <chris.friesen@windriver.com>
Depends-On: https://review.openstack.org/650288