add service dependencies to ensure reapply application triggered by unlock work
the action "host-unlock" will trigger an application reapply on
active controller and reboot the unlock controller. In simplex
system, this two processes proceed on one node and impact each
other.

when reapply the system application, the registry for docker and
a token for it were required. they should not be disabled by sm
until reapply application process was done.

So, this patch set service dependencies below to ensure sysinv
conductor can download docker images it required during
application reapplying:
1. sysinv-conductor dependent on docker-distribution
2. registry-token-server dependent on keystone

Verification:
1. verified on Simplex
2. do not impact Duplex and Multi

Closes-bug: 1822697

Change-Id: Iae4411bc64ddc8057a4b99162db6590566b3b9c2
Signed-off-by: SidneyAn <ran1.an@intel.com>